+++ EssentialFeedTests/Feed Cache/ValidateCachedFeedUseCaseTests.swift
@@ -63,45 +63,4 @@ class ValidateCachedFeedUseCaseTests: XCTestCase {
         
         return (sut, store)
     }
-    
-    // MARK: Mocking Data & Errors
-    private func mockUniqueImageFeed() -> [FeedImage] {
-        var images = [FeedImage]()
-        
-        for i in 1...3 {
-            images.append(FeedImage(id: UUID(),
-                                    description: "a description \(i)",
-                                    location: "a location \(i)",
-                                    url: URL(string: "http://an-imageURL.com?id=\(i)")!))
-        }
-        
-        return images
-    }
-    
-    private func mockUniqueFeedWithLocalRep() -> (images: [FeedImage], localRepresentation: [LocalFeedImage]) {
-        let images = mockUniqueImageFeed()
-        let localImages = images.map {
-            LocalFeedImage(id: $0.id,
-                           description: $0.description,
-                           location: $0.location,
-                           url: $0.url)
-        }
-        
-        return (images, localImages)
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 1, userInfo: nil)
-    }
-}
-
-private extension Date {
-    func adding(days: Int) -> Self {
-        return Calendar(identifier: .gregorian)
-            .date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding(seconds: TimeInterval) -> Self {
-        return self.addingTimeInterval(seconds)
-    }
 }
